feat: Update API endpoints for file upload and data retrieval
This commit updates the API endpoints for file upload and data retrieval. The '/read' endpoint has been renamed to '/post' to better reflect its purpose of accepting file uploads. Additionally, a new endpoint '/view' has been added to retrieve the data from the 'output.json' file. This improves the clarity and consistency of the API endpoints.

Also, update npm dependencies to latest stable versions

This commit updates the npm dependencies in the 'weddingUI' package to their latest stable versions. Specifically, the 'ldrs' package has been added with version 1.0.2. This ensures that the project is using the most up-to-date and secure versions of the dependencies.

Refactor code in Upload.jsx

This commit refactors the code in the 'Upload.jsx' file. It introduces a loading state using the 'isLoading' variable to indicate when the file upload is in progress. The loading state is displayed with a 'Loader' component. This improves the user experience by providing visual feedback during the file upload process.